Update dClutch successor subtree to cda37c82a
Content sync, single parent. The dclutch/ tree of this commit is
byte-identical to the live tree at cda37c82a0164ce2adc64ecc2c4f99d7be2587b5.

+++ dclutch/crates/dclutch-registry-contract/src/tests.rs
@@ -1,3 +1,5 @@
+#![allow(clippy::indexing_slicing)]
+
 extern crate std;
 
 use dclutch_core_contract::ContentId;
@@ -1325,7 +1327,7 @@ fn a_cycle_terminates_on_the_hop_bound_instead_of_running_forever() {
 
 #[test]
 fn the_walk_refuses_one_hop_past_its_own_bound() {
-    let hops: std::vec::Vec<_> = (0..=u8::from(LINEAGE_WALK_MAX_HOPS_V1))
+    let hops: std::vec::Vec<_> = (0..=LINEAGE_WALK_MAX_HOPS_V1)
         .map(|step| lineage_hop(0x11 + step, 0x12 + step))
         .collect();
     let bound = usize::from(LINEAGE_WALK_MAX_HOPS_V1);
